% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backFilt.R
\name{backFilt}
\alias{backFilt}
\title{Sampling trait estimates}
\usage{
backFilt(mMat, cMat, AMat, rMat, beta2)
}
\arguments{
\item{mMat}{N by t matrix of posterior ability estimates. Column gives the time point.}

\item{cMat}{N by t matrix of posterior variance estimates. Column gives the time point.}

\item{AMat}{N by t vector of prior ability estimates. Column gives the time point.}

\item{rMat}{N by t matrix of prior variance estimates. Column gives the time point.}

\item{beta2}{Person-level autoregressive trends}
}
\description{
After the forward filtering is complete, backwards sampling begins.
}
\examples{
backFilt()
}
\keyword{Sampling}
\keyword{backwards}
\keyword{filter}

