% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betas_fifthTry.R
\name{betas_fifthTry}
\alias{betas_fifthTry}
\title{Sampling level-one HLM parameters}
\usage{
betas_fifthTry(thet, TMat, varL1, gammas)
}
\arguments{
\item{thet}{an N by M matrix of ability estimates}

\item{TMat}{a 3 by 3 level-two variance-covariance matrix}

\item{varL1}{Level-one variance}

\item{gammas}{Level-two means}
}
\description{
This function gets person level initial abilities, person-level growth terms, and person-level autoregressive trends
}
\examples{
betas_fifthTry()
}
\keyword{HLM}
\keyword{Sampling}
\keyword{level}
\keyword{one}

