% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHLM.R
\name{getHLM}
\alias{getHLM}
\title{Getting a maximum-likelihood estimate of the HLM}
\usage{
getHLM(thet)
}
\arguments{
\item{thet}{An N by M matrix of ability estimates}
}
\description{
This convenience function gets the maximum-likelihood estimate of the HLM.
}
\examples{
medEst()
Maximum-likelihod HLM
}

