% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getZs.R
\name{getZs}
\alias{getZs}
\title{Obtaining Latent Normal Variables}
\usage{
getZs(thet, responses, ak, bk)
}
\arguments{
\item{thet}{Latent ability measures}

\item{responses}{A matrix of responses}

\item{ak}{A vector of item-level discrimination values}

\item{bk}{A vector of item-level difficulties}
}
\description{
Binary data are often augmented with latent Normally distributed variables in Bayesian Statistics.
It is easier to do statistics on these Normal latent variables (generated by this function) than on binary observed variables.
The function returns an N by I*M Matrix, where N, I and M are the number of people, items, and time points, respectively.
}
\examples{
getZs()
}
\keyword{Augmentation}
\keyword{Normal}

