% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iparmsNonInf.R
\name{iparmsNonInf}
\alias{iparmsNonInf}
\title{Non-informative item sampling}
\usage{
iparmsNonInf(thet, zMat)
}
\arguments{
\item{thet}{an N by M matrix of ability estimates}

\item{zMat}{an N by I*M matrix of latent measures}
}
\description{
Without using any prior knowledge, this function gets estimates of item traits
}
\examples{
iparmsNonInf()
}
\keyword{Sampling}
\keyword{item}
\keyword{non-informative}

