% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medEst.R
\name{medEst}
\alias{medEst}
\title{Getting the median of a chain}
\usage{
medEst(parm, low, high)
}
\arguments{
\item{parm}{the chain}

\item{low}{The lowest value of the chain you want}

\item{high}{The highest value of the chain you want}
}
\description{
This convenience function gets the median of a markov chain after tossing out the burn-in values
}

